Add PyYAML dependency.
logging.dictConfig() appears to be the most extensible method of logging
configuration as a dict can be read from any data source. YAML is used
in the Python docs' logging tutorial to create a dictionary from a
human-readable file. Similarly, that is what YAML will be used for here.

Alternative file types exist that would prevent the dependency, such as
JSON, but YAML seems to be the easiest to read and edit.