Change mongo.module export style.
The functions are defined in the body of the module, and the returned
object references those defined functions, rather than the previous
style where the functions are defined in the return object.

This makes it easier to skim and determine what is exported from a
module and makes the function definitions themselves easier to read
as the indentation is less outrageous. Additionally, this provides
safety for functions defined inside the module as functions can be
referenced by their name, rather than the mongo.module.method paradigm,
which can be overwritten.

I would like for helper functions that are exported for testing purposes
to be defined with an underscore for the method name, so that it is more
apparent they're intended to be private, i.e.:

    mongo.module.publicMethod
    mongo.module._privateMethod

But that is up for discussion.