Improving the test case design
I have added another class for integration test cases in our test suite
and moved the insert_find test to the new class. I have added a bunch of
TODOs for the test cases which should be improved once we fix the return
values for the functions and add moew functionality to them.

Fixed couple of coding convention related issues in db.py and views.py.