Fixed lookup for keys in app.config and added default environment
- Following the instructions in the README causes an exception to be thrown because
none of the environment alternatives evaluate to true and the environment
variable remains unset.

- The default config values in webapps/configs/server.py are empty strings
for ADMIN_EMAILS and LOG_FILE_PATH, so the checks for whether these keys
have been set in webapps/lib/log.py return true, even though they may not
have been configured. This causes exceptions to be thrown for attempting to
write to a non-existent log location. These checks should ignore the default case
of an empty string.