Added an option for flask to serve static files
- The MWS_FLASK_STATIC variable, when set, tells flask to serve static files
- Changed the way CONFIG_FILENAME variable is handled. Previously, it was
trying to join the current directory of the configuration handler (in
webapps/lib/conf.py) with the path specified in CONFIG_FILENAME. In the README,
the example gives an absolute filepath, which was trying to be joined with another absolute
path, which python was smart enough to just ignore. We instead just use the current working
directory of the root of the git repo to resolve files.